HYD-6523 Make requirements is broken on a mac
Make requirements is broken on OS X because a check in the make file for
version deals only with rhel and doesn't cater for usage on OS X.

This means installation of packages for development is difficult to say the
least.

In the makefile include/Makefile.distro make OS X (Darwin) appear to be EL7.

 - Add test for osname before fetching the EL version.
 - Add full path for log4tailer to ensure it can be found. This is a
   development rather than a product package

Change-Id: I008a5f7baba10ec07ccc01788aefe16149418dfa
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/22678
Tested-by: Jenkins
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Tested-by: Chroma Test User
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>