HYD-6530 Failure formatting filesystem.
Formatting of the filesystem with zfs failed because of some logging
that used inspect to fetch the code. In production systems the code
is not present and so the inspect failed causing all creations to fail.

Simply reducing the debug by removing the inspection cures the problem,
in the event of a failure a full stack will be produced anyway.

 - Remove the inspections
 - Remove entirely the debug that occurs even on the happy path

Change-Id: I32ea500431f43113b5b09aaa1ae62c95a31cf7fc
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/22691
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>