HYD-6551 Volumes disappearing that are targets prevents the manager working
It is possible for a Volume that is a Target to disappear, it shouldn't 
happen but it can happen, and today with the changes for ZFS that leads 
to IML becoming inoperable because full_volume in the target endpoint
becomes null and null is not allowed.

The problem is that is target.volume is referenced in that form the
volume will be return even it not_deleted == null, but because full_volume
is a synthesized entity it will fail if not_deleted == null. The solution
is to use the _base_manager for the initial request.

 - When creating the full_volume Use the _base_manager rather than the
   'deletable manager' used by IML

Change-Id: Ief098620c35327797b7cdb80fb74bdd26a4a1c96
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/22733
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>