HYD-6224 fake_ipmi_kvm.py hangs the first time it is run
Hooks exist within power_control.py that are called when anything is 
saved to the PowerControlDevice table, this hook will then make an rpc 
call to the PowerControlService.

However in the case of the fake_opmi_kvm.py executable their is no 
PowerControlService and so the call will hang - although I would guess
it actually times out eventually.

This is a good example of how a file that has a side-effect when you
import it can lead to unexpected behaviour.

The solution is to unhook the PowerControlDevice hook after it has been
imported in fake_opmi_kvm.py. I chose this solution over the more 
correct solution of removing the side effects due to the timing of this 
change and because fake_ipmi_kvm.py is not a part of the product.

 - Unhook register_power_device after importing PowerControlDevice in fake_opmi_kvm.py

Change-Id: I9e3020275944e970e66ee6b91f2c6a169b3f96f9
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/22763
Reviewed-by: Paul Kline <paul.everett.kline@intel.com>
Tested-by: Jenkins
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Tested-by: Chroma Test User