HYD-6451 Command Configuring Fencing cancelled alert when removing host
When removing a host from IML the remove host command completes successfully
but an additional alert is generated saying Fencing configuration was
cancelled.

The cause is the when saving a PowerControlDeviceOutlet for any reason it will
trigger a update to the pacemaker fencing config even if the pacemaker
configuration has been removed - which in this case it has.

 - Test for a pacemaker configuration that is present and not removed before
   updating the fencing configuration.

Change-Id: I027ca3bf5bb291075b47a9936366fbf03620138e
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/22780
Tested-by: Jenkins
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Tested-by: Chroma Test User
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>