HYD-6229 SQL "Fatal Errors" in install.log
Historically we moved the old alert tables to 'unused_chroma_core_'[alert] so
we could undo them, we later changed to deleting them (changed in
0024_auto__del_event__add_commandalert__add_getlnetstatejob__del_field_aler.py)

However the rename threw a fatal error (which wasn't fatal) in the logs and
people got confused. Now we are confident of the original migration we can
change it to a delete and add a migration to clean up any undeleted tables
from previous migrations.

 - Change rename to be delete in 0024_auto__del_event__add_commandale
 - Clean up renamed tables in the new migration.

Change-Id: Id8285ead286c68e17458d01382008caf08647f82
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/22781
Tested-by: Jenkins
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Tested-by: Chroma Test User
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>