HYD-6598 Improve text in IML installer repo check
Product management has asked us to improve the text of the
repo check in the IML installer based on feedback from
users. Specifically, they would like us to have the error
message include directions on how to skip the check, and
that the error may be expected when using Red Hat Satellite.

This is a simple text change to the output when one or more
enabled repos aren't contactable.

  - update text in chroma-bundles/install

Change-Id: I7a764aeb3dd8b50a72eb5bfddd39c3ceb34fac28
Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-on: http://review.whamcloud.com/22793
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>