HYD-6604 Import on ZFS needs to reimport a zpool that is currently imported read-only
When importing a zpool IML does not check that a zpool that is already 
imported is not imported readonly and so whilst returning without error
the zpool would not actually be useful to the caller.

A zpool could be imported readonly for a number of reasons beyond IMLs
control, a user may have imported it readonly, IML imports them 
readonly when it scans them and could have a bug that meant it was not
then exported, or other applications could import them readonly.

IML should check that a zpool already imported is not readonly and if
that is the case export and reimport it. After re-import IML should
again check the zpool was imported not readonly. It's not clear that
this failure mode can happen but the check is virtually free so can do
no harm.

 - Update import_ in BlockDeviceZFS to export a readonly zpool and 
   reimport it. Use recursion so we do not create a special case.
 - Update zfs_properties to force a reread of the properties, this is
   required because a exported->imported device will have changed
   properties.
 - Add tests to test the behaviour of zpools that are imported readonly.
 
Change-Id: I21f014f86dc52fdfe204afd7e675b0ce2433f627
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/22812
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>