HYD-6662 DKMS has issues meaning it is not always installed correctly
https://github.com/zfsonlinux/zfs/issues/3801 describes an issue with
the spl/zfs dkms build which means that spl/zfs may not always be
installed.

A suggested workaround is to

 dkms install spl/0.6.5.1
 dkms install zfs/0.6.5.1

to force the install and so this solution (sic) is to dkms install
when ever the agent restarts which will always include after a reboot
for a kernel update. This is generally a null event but occasionally
may cause the right thing to happen.

 - Create a method by which functions can signify they will must be
   called when the agent starts. Use a decorator called
   agent_daemon_startup_function to add the function to a list which
   the daemon can all.
 - Add some tests for the new functionality.

Change-Id: Id7d189736a451292e7e2a688a8d46b4775666025
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/23278
Tested-by: Jenkins
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>