HYD-6609 Add pool list to chroma-diagnostics
This ticket is nine of the subtasks of HYD-6608.

HYD-6610 Add zpool import to CD
HYD-6611 Add zpool status to CD
HYD-6612 Add zfs list to CD
HYD-6613 Add zfs get all to CD
HYD-6614 Add zdb to CD
HYD-6615 Capture the SPL hostid
HYD-6616 Check that the content of /etc/hostid is present and is not empty
HYD-6617 Add arc_summary.py -d to CD

Add commands to CD , enabling chroma-diagnostics to gather ZFS information
will improve the quality of the data provided to system integrators and support
engineers. Also adding the command summary.py -d which provides useful
diagnostics information for the state and health of the ARC.

Check SPL hostid , If this is  0 (zero), then the system hostid is not set
correctly, which can in turn lead to a pool being imported onto two or more
hosts simultaneously, causing data corruption.

Check that the content of /etc/hostid is not empty – it's a binary file and
the content has to be "translated" into something human readable .Really
we just want to make sure that it exists and is not empty, and report
a warning if these conditions are not met.

    - Add zpool list command to Action List
    - Add zpool import command to Action List
    - Add zpool status common to Action List
    - Add zfs list command to Action List
    - Add zfs get all command to Action List
    - Add zdb -e function to examine pools that are not imported on the current
      host
    - Capture the SPL hostid from /sys/module/spl/parameters/spl_hostid.
    - Check that the content of /etc/hostid is not empty
    - Add summary.py -d command to Action List also provides useful diagnostics
      information for the state and health of the ARC.

Change-Id: I4fb13c49538063734c31c186e694723f7a0f025f
Signed-off-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-on: http://review.whamcloud.com/22952
Tested-by: Jenkins
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Tested-by: Chroma Test User
Reviewed-by: Chris Gearing <chris.gearing@intel.com>