HYD-6626 Some valid Volumes cannot be used by Lustre
With the introduction of ZFS we introduced datasets and zvols
as Volumes within IML. When a Lustre FS is created a dataset is
created which becomes the Target Volume.  However these Volumes
themselves whilst being able to storage a Lustre filesystem target
cannot be used to create a Lustre filesystem Target.

The logic is this.

User selects a zpool as a Target and durring the mkfs.lustre process
a dataset becomes the Volume for the Lustre target. But the user
cannot select a dataset to become the Volume for a Lustre target.

To allow the GUI to only show Volumes that can be selected as the
Volume for a future Lustre target this patch adds an attribute
usable_for_lustre to the LogicalDrive class. This is true for
all LogicalDrive decendents other than ZfsPool and ZfsVol which
override the value to False.

When the rest interface requests Volumes with category unused or
usable Volumes with usable_for_lustre are not returned.

This means that whilst the dataset/zvols are shown in the TreeView
they do not appear on the Volumes page or the Create filesystem page.

 - Add usable_for_lustre to LogicalDrive
 - Add usable_for_lustre to Volume
 - Create migration for Volume
 - Add filter to remove usable_for_lustre=False for category unused or
   usable
 - Add tests to test Volume behaviour

Change-Id: Ib3192890862bc92c75c52995d398b9407c5cd1e8
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/22975
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Will Johnson <william.c.johnson@intel.com>