HYD-6520 Inconsistent volumes displayed when a zpool is detected
There is still some inconsistency between the Volumes hierarchy displayed
by IML and the actual Volumes hierarchy. This patch makes some 
improvements to the algorithms to try and improve the predictability of
the outcome.

When looking for basic changes to block devices IML 'quick' look did
not spot changes to partitions or mapper devices and so the range
of the quick look has been extended.

The device helper has also been improved to add some resiliency for
the case when a device does not respond to filestat. The first time a
device fails to respond the last good response will be used if one
exists. If the device fails again None is return, we may with experience
add more 'historical returns' before giving up.

When a change is discovered on one node it provides better results if
other associated nodes rescan devices. Previous to this patch these other
nodes where just the HA Cluster nodes, increasing this to be the sum of
HA Cluster peers and any nodes with shared volumes finds a better balance.

 - Add sysblock, mapper, disk-by-id and disk-by-path to the quicklook
   for linux devices
 - Add some additional resilience to _dev_major_minor in DeviceHelper
   so that devices the fleetingly disappear are still reported.
 - Increase the rescan nodes to include shared Volume nodes and HA cluster
   peers

Change-Id: Ib519ba26f15a3e7b5577dbf24f7e7791d4ec34ea
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/22983
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>