HYD-6631 Out-of-date OS version number in managed profile display name
The managed profile currently has display text that says it is for
EL7.2, but the next release is targetting EL7.3, so needs updating.

Now that we are back to one profile, it will be more maintainable not
to have this string to update. If/when we split into multiple profiles
again, the text that makes the most sense to differentiate them will
be dependent on what the cause of the separation is anyways.

  - Remove reference to OS version in base_managed_RH7.profile
    display string.

Change-Id: I848f34c3f17530a3de0178fd82eafa9fa58a3f3f
Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-on: http://review.whamcloud.com/23085
Tested-by: Jenkins
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Tested-by: Chroma Test User
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>