HYD-6670 Update SUPPORTED_VERSIONS in the install script
IML not longer supports EL6.X change and so this should be removed
from the install script, it is also the case that EL7.3 will be released
on the 1 Nov meaning it's version will be 1611 rather than 1610

The install script lists unsupported versions and will not list the
actual 7.3 version when it is released.

EL7.3 GA data has been changed from 10/16 to 11/16 and so the release
version will change to reflect this.

Historically we have never deleted old versions as we seem to have no
process for deciding what is not supported. The dropping of EL6 however
means the for sure EL6 is not supported because the installer will fail
during installation.

Update the SUPPORTED_VERSIONS to reflect that lack of EL6 support

 - Remove all EL6 entries from SUPPORTED_VERSIONS
 - Change 7.3.1610 to 7.3.1611 reflecting the delayed GA date of 7.3

Change-Id: I28ee1d443aa53f66b3adaee58f2be9acff5af977
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/23225
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>