HYD-6686 zpools are corrupted during test runs
The zpools are protected against being imported on multiple nodes
by the use of hostid protection. IML always exports from all
applicable nodes before it imports on another node.

However pacemaker will force the import because in a failover situation
the other node may not have cleanly unmounted. We trust pacemaker to
handle the nodes correctly and have fenced other nodes before importing.

However in the case where we direct pacemaker to start a node it will
not fence all other nodes and a potential exists for it to force an
import on a node when another node has it imported.

This patch cleans up the code and drys out the functionality to ensure
that IML will always carry out best effort to ensure a zpool has been
exported from other targets before importing it.

 - Merge MakeTargetActiveStep and MountStep to produce a common step
   that can import or start a target having ensured it is exported on
   other nodes before hand.
 - Disable the creation of new pools on each test. Disabling this feature
   makes the tests less forgiving as the zpool must remain uncorrupted
   throughout the whole test cycle.

Change-Id: I1877d0fdb6705bc644945d826550e3d12bc002a8
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/23320
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>