HYD-6654 Skip parts of stats_check() in tests when zfs devices present
Skipping just the byte write check is not sufficient due to
intermittent test failures when calculating number of files created
when writing to lustre fs with zfs-backed targets.

To avoid the test failures, files created check should also be
skipped when zfs devices are detected in the test configuration.

Currently, only the part of the check which measures bytes written is
skipped and the failure occurs when measuring the number of files
written (which differs with zfs due to the difference in use of
underlying meta data when writing to the fs).

  - Skip 'files created' and 'bytes written' checks when zfs devices
    detected in SSI or EFS test cfg
  - Remove code that skips stats_check() when zfs devices detected
    from test_filesystem_detection.py (EFS) as skip logic now present
    in stats_testcase_mixin.py

Change-Id: I190471be063ee19ea7825eb2c81ad19b96d5985e
Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-on: http://review.whamcloud.com/23275
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>