HYD-6654 Skip parts of stats_check() in tests when zfs devices present
ZFS backed targets incur extra space overhead when writing to disk (due
to differences including ZFS's use of metadata)

This results in failing tests because of unsuccessful checks in
stats_testcase_mixin.check_stats() method used in managed_failover and
filesystem_detection tests.

Relevant parts of check should be disabled until the overhead of bytes
written to ZFS devices can be reliably quantified.

Skipping just the byte write check is not sufficient due to intermittent
test failures when calculating number of files created when writing to
lustre fs with zfs-backed targets.

To avoid the test failures, the 'files created' and 'bytes total' checks
should be skipped when zfs devices are detected in the test configuration.

  - Skip 'files_created' and 'bytes_total' checks when zfs devices detected
    in SSI or EFS test cfg

Change-Id: Ia02cd15df85791d0b592ae47d390c43ef0d2f95f
Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-on: http://review.whamcloud.com/23345
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>