HYD-6356 Add debug window to failing test
Sometimes test operations start to take longer than the standard test
timeout or specified time. This begs a common question of "would we get
the correct state if we waited a bit longer", but at the same time
not wanting to hide regressions when we do have specific time
expectations.

This patch adds a new parameter to wait_until_true()
called debug_timeout, which allows you to specify an additional window
to wait and see if an expression passes for debug purposes, but an
error is still raised if it doesn't complete in the original 'timeout'.

  - add debug_timeout to wait_until_true in utility_testcase.py
  - extracted from Kelsey's original patch at
    http://review.whamcloud.com/#/c/17762/6

Change-Id: Iad60bd3662c082266f0a955a77652d3a2dc3c9da
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/23291
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>