HYD-6699 Update to Lustre build version 2.5.42.21 build 287
There was a build issue with the last version of 2.5.42.21 that
we updated to such that MPICC was missing. The Lustre team has
indicated we need to update to the properly built build #287.

Update to the Lustre version 2.5.42.21 build 287 by changing the following build parameter:

  JENKINS_BUILD_NUMBER   = 287

The following chroma-dependencies are affected:
  lustre[-client]-{el6.7,el7,sles11,sles12}
  zfs

  - Update chroma-dependencies/include/Makefile.lustre_build_meta

Change-Id: I6eb7d979d216b7dffb3d15357f3cf0144af5abf5
Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-on: http://review.whamcloud.com/23312
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>