HYD-6713 cli is broken by HYD-5449.
The change in HYD-5449 whilst useful had a side effect of turning
anything that could be parsed as a number turned into a number.

This highlights an issue when UUID are numbers (which for zfs they are)
and causes an error due to some code which is particular about receiving
a string. (config.get('targets', uuid) in _get_target_config).

Rather than change this critical code this patch places some limits on
the strings that will be parsed as json - limiting it to strings beginning
with [ or {.

 - Add test for [ or { to _codify_args
 - Make _codify_args take an args rather than a dict as an input

Change-Id: I898cb35928576a0e273936cfdb81788ae716561f
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: https://review.whamcloud.com/23405
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Joe Grund <joe.grund@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>