HYD-6863 Remove nodename from IML
IML has the concept of nodename and fqdn. The only reason for nodename 
was because pacemaker on el6 returns the nodename rather than the fqdn. 
Rather than turn the nodename into an unambiguous fqdn on the agent, 
which is actually the only safe place to really do it, nodename was 
added to the manager so that we have two methods of describing a host - 
an ambiguous nodename and a unambiguous fqdn.

The patch removes node name which is very straight forward as there is
nowhere in the code that fqdn cannot be used instead.

 - Remove nodename it serves no purpose from all code.
 - Update pacemaker so if we return to el6 nodename will be correctly 
   given as fqdn

Change-Id: I01cc8375026a13b5917943b3ed0b38dee8f1922c
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: https://review.whamcloud.com/23424
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Tested-by: Jenkins
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>