HYD-6733 test_remote_firewall_control unit tests should be removed
We should not have unit tests for RemoteFirewallControl class as this
was previously moved from production to test code. There exists the
situation where test runner executing test_remote_firewall_control.py
imports RemoteFirewallControl which in turn imports
RealRemoteOperations (which is test code that should not be imported
by unittests).

  - Remove unit/lib/test_remote_firewall_control.py unit tests

Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>
Change-Id: I4a2cc11e30eaf25bce517c73a95cc5cea4816605
Reviewed-on: https://review.whamcloud.com/23475
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>