HYD-6788 Race condition in ExportedZfsDevice
ExportedZfsDevice has a failure mode when multiple threads use it at the same
time. The original code was weak in that it only locked when the device was
actually-imported - however this means another thread that didn't lock the
device would not acquire the lock and so the original thread could export
whist the second thread is using the device - leading to failure.

It is also the case that the permanent import or export of a pool needs to
work in synchronization with the ExportedZfsDevice so that a device is not
imported and then exported by a thread running in ExportedZfsDevice.

 - Update ExportedZfsDevice to use the device for all import/export
   operations OR cases where an import/export operation might have
   taken place

Change-Id: Id6597a69a571e3ba798382ab2355b37d60186b6c
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/23611
Tested-by: Jenkins
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Tested-by: Chroma Test User