HYD-6787 IML does not discover NVMe devices
Sometimes devices, NVMe at this point, don't provide a size
entry in /sys/block/[device] and this causes the IML code
to fail in such a way that no devices are reported.

The code should be more robust so that devices that do not
conform are not reported but other devices are.

 - Add a catch to the reading of the files in /sys/block/[device]
   so that those non-conformant devices are not reported.

Change-Id: I8579946bb27268abe57f21bbb0c58eb25c14e35f
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/23646
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
(cherry picked from commit d55cd2ffc34ba726178dc29b25d102ff984c28ef)
Reviewed-on: https://review.whamcloud.com/24548