HYD-6801 ZFS issues system commands in monitor mode
The initialise_driver method which is called at ManagedNode install time but
also every time the agent starts carries out actions even in monitored mode.
The actions are to create the hostid file for zfs (if it does not exist) and
issue dkms commands to work around
https://github.com/zfsonlinux/zfs/issues/3801
namely it does dkms install for zfs/spl packages that are installed (install
does the actual build).

This actions should only occur in the case the Managed mode case and not in
the monitor mode case.

 - Make initialize_driver take managed_mode as a parameter
 - In BlockDeviceZFS initialize_driver do nothing if monitored mode.
 - Provide safe way of detected that managed mode is configure. This safe
   way should default to not in the case of an unknwon state
 - Use this safe method to detect managed mode and for completeness make
   code that detected directly use the new provider

Change-Id: I7e30811bcb8a12d05d48cb55d95c56e6d17d10bd
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/23669
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>