HYD-6807 set_volume_mounts test failure on no secondary_volume_node_id
A number of our test cases have a race where it fetches usable volumes
and then attempts to set the primary and secondary nodes on that volume.
However, the devices really have to be discovered on both hosts
separately before you can go about setting the primary and secondary on
them. This will not necessarily be instantaneous, and is thus a race
between the test case code and the product. Tests that only use one host
are not subject to this issue.

There already exists a function wait_for_shared_volumes, which is what
these tests really should be using in order to ensure the volumes are
discovered on both hosts before we try to set the primary and secondary
servers on them. These tests simply need updating to use this function.

  - Update all tests calling set_volume_mount to get its list of volumes
    from wait_for_shared_volumes, aside from those which do already.
  - Updated one instance of a test that refers to its volumes list as
    'ha_volumes' when the test only has one host, so, they are not
    actually HA volumes, which was found while looking for instances of
    this issue.

Change-Id: I303a07ca4e7237fa0487d4d296a771a4c5cd16f1
Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-on: http://review.whamcloud.com/23683
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>