HYD-6809 VolumeNode matching query does not exist
Bug exists where volume_node is not found for secondary_mount
because during UpdateManagedTargetMount target is not mounted
on the secondary host.

The problem is that a race condition exists in the trigger_plugin_update
code meaning that a trigger that occurs whilst a scan is in progress
(on the linux device) could result in a change being missed.

The background is that when a change of Linux resource is discovered by the
manager on a Node the manager will trigger a scan all all other related nodes.

Related nodes are the superset of the HA pairs of the Node and any Nodes that
share Disks with the Node.

This behaviour is required because (as an example) if a dataset is created on
Node A then Node B cannot see the dataset until it imports the zpool - how it
can't import zpools just to see if something has changed.  The trigger means
that with Node A having seen the change it will cause all other related nodes
to do a full scan so they sync the changes.

The problem is that the 'trigger flag' was being reset after the scan in the
linux agent plugin later than before.

Linux plugin   -- Scan starts
Linux plugin   -- Device X Scanned
Another Node   -- Device X changes
Another Thread -- Trigger Flag Set
Linux plugin   -- Scan ends
Linux plugin   -- Trigger Flag Reset

(Trigger flag left unset)

The changes to Device X will never not be reported because the flag was reset after
the change and trigger had occurred.

If the trigger flag is reset before the scan then changes will never be
missed

Linux plugin   -- Trigger Flag Reset
Linux plugin   -- Scan starts
Linux plugin   -- Device X Scanned
Another Node   -- Device X changes
Another Thread -- Trigger Flag Set
Linux plugin   -- Scan ends

(Trigger flag left set and so another scan will occur and report the changes
to Device X)

 - Change the clearing of the trigger_plugin_update flag to ensure
   no changes can be missed.

Change-Id: I2ca77380975cf46e99f264ae289f30bca73d5e02
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/23761
Tested-by: Jenkins
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Tested-by: Chroma Test User