HYD-6786 Fix fqdn_matches failure in SSI tests
socket.getfqdn() seems to be returning an empty string and failing
checks within test_host_contact.py

Add retries to the test code to allow this flakey behaviour to be
worked around. By adding this workaround to the test code we are
really behaving like a user who would in truth wait for the red-box
to turn into a green-box.

  - Retry the posts to /api/test_host until a good validation occurs
    for all hosts.
  - After the standard TEST_TIMEOUT exit with an exception.

Change-Id: Iaad6ea5441af8687eaf61281521f394b56bfb3ba
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/23785
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>