HYD-6849 Make check_stats() work for all backend filesystem types
EFS is failing intermittently due to check_stats() errors inside
test_filesystem_detection.py. In previous patches, parts of
check_stats have been skipped when zfs backend_filesystems have been
detected in the test cluster config (HYD-6654).

check_stats() aims to verify expected reporting of fs changes in IML
API. As it is functional verification, exact measurement is not
necessary as testing accuracy of Lustre stats is not within the remit
of IML. Exact changes in stats can differ dependent on backend fs.

Having a more robust mechanism to wait for up-to-date statistics may
offer a better solution rather than just skipping check_stats() for
EFS tests as has been done in the past.

In an attempt to improve the reliability of the test, repeatedly
request and compare stats until a timeout occurs.

In order to generalise the test to work with different backend fs
types, generate a 100MB random file in tmp, copy that to the
filesystem, unmount, check the space on the filesystem has changed by
~100MB (+/- 10%). Then mount, read the file back by comparing it to
the random file and check no differences.  In summary check the
reported filesystem free space is within 10% of what it is expected
to be after the write.

It should be noted that if reduce_fn is set to 'sum' in url query
string any supplied metric name will be given a default value of 0.0
and returned regardless of whether a value is present in the
MetricStore.

  - Validate all expected metrics are being returned from API
    before commencing checks
  - Retry fetching and testing metrics until TEST_TIMEOUT expires
  - Generate random file and copy to filesystem
  - Check free space on filesystem through API
  - Compare file contents copied to filesystem with original file
  - Make use of retry utilities to DRY-out and make test more robust
  - Fetch-help to catch any validation failures during new checks
  - Remove some redundant code in test_filesystem_detection.py
  - Move code exclusively used in test_filesystem_dne.py into the
    file from stats_testcase_mixin.py
  - Add retry for test_filesystem_dne.py _check_stats() and
    remove errorprone sleep()

Change-Id: I2bcdf3c71afb3638a7a120b285b33611b3cf3e93
Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-on: https://review.whamcloud.com/23875
Tested-by: Jenkins
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Tested-by: Chroma Test User
Reviewed-by: Chris Gearing <chris.gearing@intel.com>