HYD-6845 Test failure: RpcError - missing parent resource
For reasons that we do not understand a KeyError occurs when adding a
parent in very rare circumstances.

It may be connected with the order children and parents are added.

It may be connected with the program being restarted meaning the DB
has records in it that the in memory local_id_to_global_id does not.

It may be that because the DB changes are transaction based and the
local_id_to_global_id are not the two get out of sync.

It may be thread related.

At this point we do not know and so this patch created this is the worst kind
of hack, no excuses but short of not delivering the the product this is the
best we can come up with. It is possibly an ordering thing that the parent has
not been created at the point the child is created, possibly something else.

If the former then we believe that the code will on the next update have the
parent and so update the child. If not then items in the tree view may show
items as leafs that are actually branches and give the user an option they
should not use. For example left them select a disk when it has a partition

Only time will tell but if this patch landed and you are reading this then we
never came up with a proper solution in the time available.

The hack is to not create parents when the local_id_to_global_id record is not
available.

  - Return from session_resource_add_parent without creating the parent link if
    the session_resource_add_parent

Change-Id: Ide57dd0544de5ee03e4fb2590264d0be1d5b0f76
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/23895
Tested-by: Jenkins
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Tested-by: Chroma Test User